Refactor summary/cli utilities and add portfolio tests
